pyproject.toml
[project]
name = "report-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "yt-dlp>=2025.12.8",
]
